Move Compiler to top level type
