+++ core/shared/src/main/scala/fs2/internal/ScopedResource.scala
@@ -25,7 +25,8 @@ import cats.MonadError
 import cats.effect.Resource
 import cats.effect.concurrent.Ref
 import cats.implicits._
-import fs2.Scope
+
+import fs2.{Compiler, Scope}
 
 /**
   * Represents a resource acquired during stream interpretation.
@@ -125,7 +126,7 @@ private[internal] object ScopedResource {
 
   private[this] val initial = State(open = true, finalizer = None, leases = 0)
 
-  def create[F[_]](implicit F: CompilationTarget[F]): F[ScopedResource[F]] =
+  def create[F[_]](implicit F: Compiler.Target[F]): F[ScopedResource[F]] =
     for {
       state <- F.ref[State[F]](initial)
       token <- Token[F]
